% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.top5.R
\name{data.top5}
\alias{data.top5}
\title{Cálcular las cinco especies más abundantes por proyecto}
\usage{
data.top5(data, proyecto)
}
\arguments{
\item{data}{objeto con la base de datos}

\item{proyecto}{objeto con el código del proyecto (BH_PE, BH_LT)}

\item{periodo}{fechas del periodo que abarca el análisis con formato aaaa-mm-dd hh:mm:ss}
}
\value{
los datos acomodados para la elaboración del gráfico
}
\description{
Calcula las cinco especies más abundantes por proyecto listas en la estructura necesaria para graficar
}
\author{
Biol. Michell Romero - Coordinador del área de Monitoreo de aves TCT 🦉
}
