% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.endemismo.R
\name{graf.endemismo}
\alias{graf.endemismo}
\title{Gráfico de especies por categorias de endemismo}
\usage{
graf.endemismo(data, paleta_colores)
}
\arguments{
\item{data}{objeto con la hoja que contiene los datos de spp. por categoría de endemismo (únicas)}

\item{paleta_colores}{vector con la paleta de colores asignada a cada categoría de endemismo}
}
\value{
gráfico de pastel con el número de spp. por categoria de endemismo
}
\description{
Gráfico de pie o pastel para el número de especies por categoría de endemismo mediante la información resumida de "resumen.taxo"
}
\author{
Biol. Michell Romero - Coordinador del área de Monitoreo de aves TCT 🦉
}
