% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.residencia.R
\name{graf.residencia}
\alias{graf.residencia}
\title{Gráfico de especies por categorias de residencia}
\usage{
graf.residencia(data, paleta_colores)
}
\arguments{
\item{data}{objeto con la hoja que contiene los datos de spp. por residencia}

\item{paleta_colores}{vector con la paleta de colores asignada a cada categoría de residencia}
}
\value{
gráfico de pastel con el número de spp por categoria de residencia
}
\description{
Gráfico de pie o pastel para el número de especies por categoría de residencia mediante la información resumida de "resumen.taxo"
}
\author{
Biol. Michell Romero - Coordinador del área de Monitoreo de aves TCT 🦉
}
