% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.top5.R
\name{graf.top5}
\alias{graf.top5}
\title{Gráfico de las top 5 especies más abundantes}
\usage{
graf.top5(data, colores)
}
\arguments{
\item{data}{objeto con la base de datos generada con la función data.top5}

\item{colores}{vector con la paleta de colores asignada a cada código de sp.}
}
\value{
gráfico de barras de las cinco especies más abundantes
}
\description{
Gráfico de barras para las cinco especies más abundantes
}
\author{
Biol. Michell Romero - Coordinador del área de Monitoreo de aves TCT 🦉
}
