% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumen.taxo.R
\name{resumen_taxo}
\alias{resumen_taxo}
\title{Resumen de taxonomia por categorias de riesgo, estacionalidad y endemismos}
\usage{
resumen_taxo(taxo_df)
}
\arguments{
\item{taxo_df}{listado de especies con sus respectivas categorias}
}
\value{
Resumen del número de especies por cada categoría de riesgo, endemismo y gremios
}
\description{
Describe y resume el número de especies protegidas, por categorías de estacionalidad y endemismos.
(minúsculas, sin acentos, sin espacios).
}
\author{
Biol. Michell Romero - Coordinador del área de Monitoreo de aves TCT 🦉
}
