% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_totales.R
\name{spp.totales}
\alias{spp.totales}
\title{Cálculo de especies totales en todos los métodos y proyectos.}
\usage{
spp.totales(lista_rutas, hojas)
}
\arguments{
\item{lista_rutas}{ruta a la carpeta donde se encuentran las distintas bases de datos (proyectos) en .xlsx}

\item{hojas}{hojas que contienen la columna "Especie" para extraer taxonomía.}
}
\value{
Un archivo .xlsx con el inventario de spp. totales
}
\description{
Esta función selecciona la columna Especie de todas las hojas y archivos asignados para regresar un .xlsx con spp. únicas.
}
\author{
Biol. Michell Romero - Coordinador del área de Monitoreo de aves TCT 🦉
}
